toiurnament entry logics and resolves implemented

+++ database/xayaplayer.hpp
@@ -325,7 +325,7 @@ public:
 private:
    
   /*Helper utility function used in perstige calculations*/
-  float GetFighterPercentageFromQuality(uint32_t quality, std::vector<FighterTable::Handle>& fighters);
+  uint32_t GetFighterPercentageFromQuality(uint32_t quality, std::vector<FighterTable::Handle>& fighters);
   
   /*We are not writing those in the database, so we need to create them on each 
   time we load account or create fresh*/
